refactor: move Terminal.hpp and Terminal.cpp to Terminal directory
