+++ src/EditorConfig.cpp
@@ -1,6 +1,6 @@
 #include "EditorConfig.hpp"
 
-#include "Terminal.hpp"
+#include "Terminal/Terminal.hpp"
 #include <iostream>
 #include <system_error>
